Handle blob path discrepancies/duplicates
The JSON blob (parts of which are generated by different users
for each platform) has discrepancies in the paths/working directories
for each of the platform specific targets. This change allows the
generator to generate targets names that contain paths relative to the
nodesrc root directory.

Keep track of targets already written to CMakelists to prevent duplicates.

Move the analysis generation from the writer to the generator script.