analyse: treat header only libs as interface libs
The 'v8_headers' target was introduced as a header-only target, and this
causes issues when creating such a target in CMake.  Treating these as
interface libraries avoids this problem.

Additionally, do not generate a link directive for interface libraries
that are actually custom targets, as this causes issues with CMake.